Add ReverseAfterEndAnimation. Animations moved to separate namespace
